--- Project5/Pixel.cpp
@@ -1,1 +1,0 @@
-#include "Pixel.h"
