--- Project5/Slider.cpp
@@ -1,1 +1,0 @@
-#include "Slider.h"
