--- Project5/Timer.cpp
@@ -1,1 +1,0 @@
-#include "Timer.h"
