+++ Project5/src/library/Pixel.cpp
@@ -1,0 +1,1 @@
+#include <library/Pixel.h>
