+++ Project5/src/library/Slider.cpp
@@ -1,0 +1,1 @@
+#include <library/Slider.h>
